EditProfileFragment.kt save almost done (Kotlin Reflection required and GSON dependency for JSON and DataClass parse to Map)

+++ app/src/main/java/com/feryaeldev/djexperience/data/models/User.kt
@@ -9,7 +9,7 @@ data class User(
     var email: String? = "",
     var country: String? = "",
     var category: String? = "",
-    var age: Int? = 0,
+    var age: Long? = 0,
     var website: String? = "",
     var following: ArrayList<String>? = arrayListOf()
 )
